Cast x_beg, x_end, y_beg, y_end from int to size_t

+++ src/receptor.cpp
@@ -150,8 +150,8 @@ int receptor::populate(const scoring_function& sf, const vector<size_t>& xs, con
 		const float dydx_ub = sqrt(dydx_sqr_ub);
 		const float y_lb = a.coord[1] - dydx_ub;
 		const float y_ub = a.coord[1] + dydx_ub;
-		const size_t y_beg = y_lb > corner0[1] ? (y_lb < corner1[1] ? static_cast<int>((y_lb - corner0[1]) * granularity_inverse) : num_y_probes) : 0;
-		const size_t y_end = y_ub > corner0[1] ? (y_ub < corner1[1] ? static_cast<int>((y_ub - corner0[1]) * granularity_inverse) + 1 : num_y_probes) : 0;
+		const size_t y_beg = y_lb > corner0[1] ? (y_lb < corner1[1] ? static_cast<size_t>((y_lb - corner0[1]) * granularity_inverse) : num_y_probes) : 0;
+		const size_t y_end = y_ub > corner0[1] ? (y_ub < corner1[1] ? static_cast<size_t>((y_ub - corner0[1]) * granularity_inverse) + 1 : num_y_probes) : 0;
 		const vector<size_t>& p = p_offset[a.xs];
 		size_t zy_offset = z_offset + num_x_probes * y_beg;
 		float dy = corner0[1] + granularity * y_beg - a.coord[1];
@@ -163,8 +163,8 @@ int receptor::populate(const scoring_function& sf, const vector<size_t>& xs, con
 			const float dx_ub = sqrt(dx_sqr_ub);
 			const float x_lb = a.coord[0] - dx_ub;
 			const float x_ub = a.coord[0] + dx_ub;
-			const size_t x_beg = x_lb > corner0[0] ? (x_lb < corner1[0] ? static_cast<int>((x_lb - corner0[0]) * granularity_inverse) : num_x_probes) : 0;
-			const size_t x_end = x_ub > corner0[0] ? (x_ub < corner1[0] ? static_cast<int>((x_ub - corner0[0]) * granularity_inverse) + 1 : num_x_probes) : 0;
+			const size_t x_beg = x_lb > corner0[0] ? (x_lb < corner1[0] ? static_cast<size_t>((x_lb - corner0[0]) * granularity_inverse) : num_x_probes) : 0;
+			const size_t x_end = x_ub > corner0[0] ? (x_ub < corner1[0] ? static_cast<size_t>((x_ub - corner0[0]) * granularity_inverse) + 1 : num_x_probes) : 0;
 			const float dzdy_sqr = dz_sqr + dy_sqr;
 			size_t zyx_offset = zy_offset + x_beg;
 			float dx = corner0[0] + granularity * x_beg - a.coord[0];
